CircleAndLinesPainter is no longer private (used for current/past visits); moved to utils
