+++ lib/utils/current_visit_data.dart
@@ -6,3 +6,5 @@ class CurrentVisitData {
   final String timeOfEvent;
   final String eventDescription;
 }
+
+// TODO: Do we need a PastVisitData? Or is this not necessary? - DY
